+++ tests/test_starrail.py
@@ -21,4 +21,5 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import starrail
+import pytest
+from starrail import HonkaiStarRail