+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # Change Log
 
+### Version 0.0.2 - [4/29/2023]
+Added commandline support for the following operations (**UNSTABLE**):
+
+1. Configure `starrail` (only once after download):
+```shell
+$ starrail configure
+```
+2. Starting Honkai: Star Rail from the commandline
+```shell
+$ starrail start
+```
+3. Terminating Honkai: Star Rail from the commandline (started from `starrail`)
+```shell
+$ starrail stop
+```
+
 ### Version 0.0.1 - [4/26/2023]
 Initial Release